+++ Package.swift
@@ -3,7 +3,7 @@ import PackageDescription
 let package = Package(
     name: "Sourcery",
     targets: [
-        Target(name: "sourcery", dependencies: []),
+        Target(name: "sourcery", dependencies: [])
     ],
     dependencies: [
         // https://github.com/kylef/Stencil/pull/84
@@ -13,6 +13,6 @@ let package = Package(
         .Package(url: "https://github.com/vknabel/PathKit.git", majorVersion: 0, minor: 7),
         .Package(url: "https://github.com/jpsim/SourceKitten.git", majorVersion: 0, minor: 15),
         .Package(url: "https://github.com/vknabel/SwiftTryCatch.git", majorVersion: 1, minor: 1),
-        .Package(url: "https://github.com/krzysztofzablocki/KZFileWatchers.git", majorVersion: 1, minor: 0),
+        .Package(url: "https://github.com/krzysztofzablocki/KZFileWatchers.git", majorVersion: 1, minor: 0)
     ]
 )